+++ src/abstractions/Extensions.cs
@@ -1,8 +1,106 @@
 ﻿using System;
+using System.IO;
+using System.Linq;
+using System.Text;
+using NeoExpress.Abstractions.Models;
+using Newtonsoft.Json;
 
 namespace NeoExpress.Abstractions
 {
-    public class Extensions
+    public static class Extensions
     {
+        public static string ToHexString(this byte[] value, bool reverse = false)
+        {
+            var sb = new StringBuilder();
+
+            if (reverse)
+            {
+                for (int i = value.Length - 1; i >= 0; i--)
+                {
+                    sb.AppendFormat("{0:x2}", value[i]);
+                }
+            }
+            else
+            {
+                for (int i = 0; i < value.Length; i++)
+                {
+                    sb.AppendFormat("{0:x2}", value[i]);
+                }
+            }
+            return sb.ToString();
+        }
+
+        public static byte[] ToByteArray(this string value)
+        {
+            if (value == null || value.Length == 0)
+                return new byte[0];
+            if (value.Length % 2 == 1)
+                throw new FormatException();
+            byte[] result = new byte[value.Length / 2];
+            for (int i = 0; i < result.Length; i++)
+                result[i] = byte.Parse(value.Substring(i * 2, 2), System.Globalization.NumberStyles.AllowHexSpecifier);
+            return result;
+        }
+
+        public static void Save(this ExpressChain chain, string fileName)
+        {
+            var serializer = new JsonSerializer();
+            using (var stream = File.Open(fileName, FileMode.Create, FileAccess.Write))
+            using (var writer = new JsonTextWriter(new StreamWriter(stream)) { Formatting = Formatting.Indented })
+            {
+                serializer.Serialize(writer, chain);
+            }
+        }
+
+        public static bool IsReservedName(this ExpressChain chain, string name)
+        {
+            if ("genesis".Equals(name, StringComparison.InvariantCultureIgnoreCase))
+                return true;
+
+            foreach (var node in chain.ConsensusNodes)
+            {
+                if (string.Equals(name, node.Wallet.Name, StringComparison.InvariantCultureIgnoreCase))
+                    return true;
+            }
+
+            return false;
+        }
+
+        public static bool NameEquals(this ExpressContract contract, string name) =>
+            string.Equals(contract.Name, name, StringComparison.InvariantCultureIgnoreCase);
+
+        public static bool NameEquals(this ExpressWallet wallet, string name) =>
+            string.Equals(wallet.Name, name, StringComparison.InvariantCultureIgnoreCase);
+
+        public static ExpressWallet GetWallet(this ExpressChain chain, string name) =>
+            (chain.Wallets ?? Enumerable.Empty<ExpressWallet>())
+                .SingleOrDefault(w => w.NameEquals(name));
+
+        public static ExpressWalletAccount? GetAccount(this ExpressChain chain, string name)
+        {
+            if (chain.Wallets != null)
+            {
+                var wallet = chain.Wallets.SingleOrDefault(w => w.NameEquals(name));
+                if (wallet != null)
+                {
+                    return wallet.DefaultAccount;
+                }
+            }
+
+            var node = chain.ConsensusNodes.SingleOrDefault(n => n.Wallet.NameEquals(name));
+            if (node != null)
+            {
+                return node.Wallet.DefaultAccount;
+            }
+
+            if ("genesis".Equals(name, StringComparison.InvariantCultureIgnoreCase))
+            {
+                return chain.ConsensusNodes
+                    .Select(n => n.Wallet.Accounts.Single(a => a.Label == "MultiSigContract"))
+                    .FirstOrDefault();
+            }
+
+            return null;
+        }
     }
 }
