+++ src/neo-express/Commands/ClaimCommand.cs
@@ -1,4 +1,5 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
+using NeoExpress.Abstractions;
 using NeoExpress.Neo2;
 using System;
 using System.Threading.Tasks;
