+++ src/neo-express/Commands/WalletCommand.Create.cs
@@ -1,4 +1,5 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
+using NeoExpress.Abstractions;
 using NeoExpress.Abstractions.Models;
 using System;
 using System.Collections.Generic;
