+++ src/neo-express/Commands/WalletCommand.Delete.cs
@@ -1,4 +1,5 @@
 ﻿using McMaster.Extensions.CommandLineUtils;
+using NeoExpress.Abstractions;
 using System;
 using System.ComponentModel.DataAnnotations;
