+++ src/neo-express/Commands/WalletCommand.Export.cs
@@ -1,7 +1,8 @@
-﻿using McMaster.Extensions.CommandLineUtils;
-using System;
+﻿using System;
 using System.ComponentModel.DataAnnotations;
 using System.IO;
+using McMaster.Extensions.CommandLineUtils;
+using NeoExpress.Abstractions;
 
 namespace NeoExpress.Commands
 {
